Move Metadata -> ImageResult.Metadata.
