+++ coil-singleton/api/coil-singleton.api
@@ -29,7 +29,7 @@ public final class coil/ImageViews {
 	public static synthetic fun load$default (Landroid/widget/ImageView;Lokhttp3/HttpUrl;Lcoil/ImageLoader;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcoil/request/Disposable;
 	public static final synthetic fun loadAny (Landroid/widget/ImageView;Ljava/lang/Object;Lcoil/ImageLoader;Lkotlin/jvm/functions/Function1;)Lcoil/request/Disposable;
 	public static synthetic fun loadAny$default (Landroid/widget/ImageView;Ljava/lang/Object;Lcoil/ImageLoader;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcoil/request/Disposable;
-	public static final fun metadata (Landroid/widget/ImageView;)Lcoil/request/Metadata;
+	public static final fun metadata (Landroid/widget/ImageView;)Lcoil/request/ImageResult$Metadata;
 }
 
 public final class coil/api/ImageViews {
